+++ libnavui-maps/src/main/java/com/mapbox/navigation/ui/maps/internal/route/line/MapboxRouteLineUtils.kt
@@ -1104,7 +1104,7 @@ internal object MapboxRouteLineUtils {
             RouteLayerConstants.WAYPOINT_SOURCE_ID,
             options.tolerance,
             useLineMetrics = false,
-            options.shareLineGeometrySources
+            enableSharedCache = false
         )
         addSource(
             style,