+++ CHANGELOG.md
@@ -31,6 +31,7 @@ This release depends on, and has been tested with, the following Mapbox dependen
 
 - Added experimental option to MapboxRouteLineOptions to enable the shared cache in the Maps SDK. [#6680](https://github.com/mapbox/mapbox-navigation-android/pull/6680)
 - Added experimental `MapboxRouteLineOptions#enableSharedCache` option to enable route line's GeoJson source data sharing between multiple instances of the map. [#6680](https://github.com/mapbox/mapbox-navigation-android/pull/6680)
+- Added experimental `MapboxRouteLineOptions#shareLineGeometrySources` option to enable route line's GeoJson source data sharing between multiple instances of the map. [#6680](https://github.com/mapbox/mapbox-navigation-android/pull/6680)
 
 ## Mapbox Navigation SDK 2.10.0-beta.2 - 01 December, 2022
 ### Changelog